flights/requirements.txt
Flask==3.0.3
flasgger==0.9.7.1
Flask-Cors==5.0.0

# OpenTelemetry dependencies
opentelemetry-sdk
opentelemetry-api
opentelemetry-distro[otlp]
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-requests
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp
opentelemetry-semantic-conventions
opentelemetry-instrumentation-system-metrics
opentelemetry-instrumentation-logging
